moved each function into their own directory.
