+++ tests/index.test.ts
@@ -3,7 +3,7 @@ import "../src/index";
 import { assert, expect } from "chai";
 import { describe } from "mocha";
 
-import { assertErrorType, Exception_FindReplaceIllegalAction } from "../src/customErrors";
+import { assertErrorType, Exception_FindReplaceIllegalAction } from "../src/utils/customErrors";
 import { myArr, Score } from "./testData";
 
 const ASSERTIONS = {
